refactor(terminal): split terminal helpers into modules
Break the ccusage-terminal crate out of the single lib.rs file while keeping the public exports unchanged.

Move table rendering, terminal styling, title rendering, width handling, and terminal width discovery into focused private modules so future terminal work has clearer ownership.

The snapshot outputs are unchanged; their files move only because the tests now live beside the extracted modules.
