+++ rust/crates/ccusage-terminal/src/width.rs
@@ -1,0 +1,44 @@
+pub(crate) fn visible_width(value: &str) -> usize {
+    let bytes = value.as_bytes();
+    let mut width = 0;
+    let mut index = 0;
+    while index < bytes.len() {
+        if bytes[index] == 0x1b {
+            index += 1;
+            if index < bytes.len() && bytes[index] == b'[' {
+                index += 1;
+                while index < bytes.len() && !(bytes[index] as char).is_ascii_alphabetic() {
+                    index += 1;
+                }
+                index += usize::from(index < bytes.len());
+            }
+            continue;
+        }
+        let Some(ch) = value[index..].chars().next() else {
+            break;
+        };
+        width += char_display_width(ch);
+        index += ch.len_utf8();
+    }
+    width
+}
+
+pub(crate) fn contains_ansi(value: &str) -> bool {
+    value.as_bytes().contains(&0x1b)
+}
+
+pub(crate) fn char_display_width(ch: char) -> usize {
+    if ch.is_ascii() {
+        1
+    } else {
+        2
+    }
+}
+
+pub(crate) fn visible_width_max_line(value: &str) -> usize {
+    value.lines().map(visible_width).max().unwrap_or_default()
+}
+
+pub(crate) fn visible_width_sum(value: &str) -> usize {
+    value.lines().map(visible_width).sum()
+}
